+++ app/models/surveys/answer_type.rb
@@ -1,0 +1,3 @@
+class AnswerType < ActiveRecord::Base
+  has_many :questions
+end