+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog for OpenPPRN
 
-## Backend
+## R1 - Social Functionality
+### Enhancements
+- Discussion forum added
+- Users can set up their social profile
+- Users can upload profile photos
+- Geocoding support implemented
+- Users can see other participants on a map
+- Admin dashboard refactored
 
-### backend-0.0.0
+## R0.1 - Admin Profile Fixes and Testing
 
-#### Enhancements
+## R0 - Surveys & Owners/Admins
 
+### Enhancements
+- Add documentation
+- Add Survey functionality
 - Add Travis-CI testing badge that shows the continuous integration test status
   - Based on Travis-CI guide here: http://docs.travis-ci.com/user/using-postgresql/#Using-PostgreSQL-in-your-Builds